+++ source/extensions/lightspeed.trex.asset_replacements.core.shared/lightspeed/trex/asset_replacements/core/shared/__init__.py
@@ -15,6 +15,18 @@
 * limitations under the License.
 """
 
-__all__ = ["Setup"]
+__all__ = [
+    "CachedReplacementSkeletons",
+    "Setup",
+    "SkeletonAutoRemappingError",
+    "SkeletonReplacementBinding",
+    "SkeletonDefinitionError",
+]
 
 from .setup import Setup
+from .skeleton import (
+    CachedReplacementSkeletons,
+    SkeletonAutoRemappingError,
+    SkeletonDefinitionError,
+    SkeletonReplacementBinding,
+)