+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/omni/flux/stage_manager/plugin/interaction/usd/all_skeletons.py
@@ -1,0 +1,49 @@
+"""
+* SPDX-FileCopyrightText: Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+* SPDX-License-Identifier: Apache-2.0
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+* https://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+"""
+
+from omni.flux.stage_manager.factory.plugins import StageManagerFilterPlugin as _StageManagerFilterPlugin
+from omni.flux.stage_manager.factory.plugins import StageManagerTreePlugin as _StageManagerTreePlugin
+
+from .base import StageManagerUSDInteractionPlugin as _StageManagerUSDInteractionPlugin
+
+
+class AllSkeletonsInteractionPlugin(_StageManagerUSDInteractionPlugin):
+    display_name: str = "Skeletons"
+    tooltip: str = "View all the Skeletons and their bound Prims"
+
+    internal_context_filters: list[_StageManagerFilterPlugin] = [
+        {"name": "SkeletonPrimsFilterPlugin"},
+    ]
+    tree: _StageManagerTreePlugin = {"name": "SkeletonGroupsTreePlugin"}
+
+    compatible_trees: list[str] = ["SkeletonGroupsTreePlugin", "PrimGroupsTreePlugin", "VirtualGroupsTreePlugin"]
+    compatible_filters: list[str] = [
+        "IgnorePrimsFilterPlugin",
+        "IsCaptureFilterPlugin",
+        "SkeletonPrimsFilterPlugin",
+        "OmniPrimsFilterPlugin",
+        "SearchFilterPlugin",
+    ]
+    # TODO StageManager: We have LSS plugin names in the flux ext because of this system
+    compatible_widgets: list[str] = [
+        "PrimTreeWidgetPlugin",
+        "IsVisibleActionWidgetPlugin",
+        "IsCaptureStateWidgetPlugin",
+        "RemapSkeletonActionWidgetPlugin",
+        "RemapSkeletonInfoWidgetPlugin",
+        "FocusInViewportActionWidgetPlugin",
+    ]