+++ source/extensions/omni.flux.stage_manager.plugin.tree.usd/omni/flux/stage_manager/plugin/tree/usd/extension.py
@@ -22,12 +22,13 @@ from omni.flux.stage_manager.factory import get_instance as _get_factory_instanc
 
 from .light_groups import LightGroupsTreePlugin as _LightGroupsTreePlugin
 from .prim_groups import PrimGroupsTreePlugin as _PrimGroupsTreePlugin
+from .skeleton_groups import SkeletonGroupsTreePlugin as _SkeletonGroupsTreePlugin
 from .virtual_groups import VirtualGroupsTreePlugin as _VirtualGroupsTreePlugin
 
 
 class StageManagerUSDTreePluginsExtension(omni.ext.IExt):
 
-    _PLUGINS = [_LightGroupsTreePlugin, _PrimGroupsTreePlugin, _VirtualGroupsTreePlugin]
+    _PLUGINS = [_LightGroupsTreePlugin, _PrimGroupsTreePlugin, _VirtualGroupsTreePlugin, _SkeletonGroupsTreePlugin]
 
     def on_startup(self, _):
         carb.log_info("[omni.flux.stage_manager.plugin.tree.usd] Startup")