+++ source/extensions/lightspeed.trex.app.resources/docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.10.0]
+## Added
+- Stage Manager: Added a skeleton interaction tab
+- Add skeleton related icons
+
 ## [1.9.1]
 ## Fixed
 - conform disclosure icon meaning so down arrow means expanded