+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.11.0]
+## Added
+- Added a skeleton interaction tab
+
 ## [1.10.1]
 ### Fixed
 - Fixed tests flakiness